rm DemographicSurveyEnum
